fix: search 페이지 서버컴포넌트로 변경 (#75)
